-Tina's script and quest changes:
If leadership and loyalty is good enough Tina in Drassen clothes factory in C13 not only gives the player information, but also the unique Bloodcat Backpack (# 3026) with some special attributes/bonuses in case the player gives her the needed silk (Parachute # 3001).

(A homage to Ambermoon (Nestraks Zauberstab) and all other great RPGs of the past!)


Credit and thanks for Tina's voice records goes to Spookygirl!